+++ src/main/java/com/nedjar/algo/hachage/HacheurSedgewick.java
@@ -5,7 +5,7 @@ import com.nedjar.dataStructure.Tuple;
 public class HacheurSedgewick implements Hacheur {
 
 	@Override
-	public long hache(Tuple key) {
+	public int hache(Tuple key) {
 		int h = 0;
 		int a = 127;
 		for (byte i = 0; i < key.size(); i++) {
